+++ frontend/src/components/settings/types.ts
@@ -43,12 +43,15 @@ export type SectionId =
     | 'api-tokens'
     | 'registries'
     | 'labels'
-    | 'system'
+    | 'host-alerts'
+    | 'docker-storage'
+    | 'fleet-mesh'
     | 'notifications'
     | 'webhooks'
     | 'security'
     | 'cloud-backup'
     | 'developer'
+    | 'data-retention'
     | 'nodes'
     | 'app-store'
     | 'notification-routing'
